Split up gui_utils module into qt.utils and io_utils
