+++ geoAssembler/qt/subwidgets.py
@@ -10,10 +10,10 @@ from PyQt5 import uic
 from pyqtgraph.Qt import (QtCore, QtGui, QtWidgets)
 
 from .objects import (CircleShape, DetectorHelper, SquareShape, warning)
+from .utils import get_icon
 
 from ..defaults import DefaultGeometryConfig as Defaults
-from ..gui_utils import (get_icon,
-                         read_geometry, write_geometry)
+from ..io_utils import read_geometry, write_geometry
 
 
 Slot = QtCore.pyqtSlot
